+++ tap_klaviyo/__init__.py
@@ -5,16 +5,19 @@ import os
 import sys
 import singer
 from singer import metadata
-from tap_klaviyo.utils import get_incremental_pull, get_full_pulls, get_all_pages
+from tap_klaviyo.utils import get_incremental_pull, get_full_pulls, get_all_using_next
 
 LOGGER = singer.get_logger()
 
+API_VERSION = "2024-10-15"
+
+# For stream global_exclusions, data related to suppressed users can be found in the /api/profiles endpoint
 ENDPOINTS = {
-    'global_exclusions': 'https://a.klaviyo.com/api/v1/people/exclusions',
-    'lists': 'https://a.klaviyo.com/api/v1/lists',
-    'metrics': 'https://a.klaviyo.com/api/v1/metrics',
-    'metric': 'https://a.klaviyo.com/api/v1/metric/',
-    'campaigns': 'https://a.klaviyo.com/api/v1/campaigns'
+    'global_exclusions': 'https://a.klaviyo.com/api/profiles',
+    'lists': 'https://a.klaviyo.com/api/lists',
+    'metrics': 'https://a.klaviyo.com/api/metrics',
+    'events': 'https://a.klaviyo.com/api/events',
+    'campaigns': 'https://a.klaviyo.com/api/campaigns'
 }
 
 EVENT_MAPPINGS = {
@@ -22,20 +25,21 @@ EVENT_MAPPINGS = {
     "Clicked Email": "click",
     "Opened Email": "open",
     "Bounced Email": "bounce",
-    "Unsubscribed": "unsubscribe",
+    "Unsubscribed from Email Marketing": "unsubscribe",
     "Marked Email as Spam": "mark_as_spam",
     "Unsubscribed from List": "unsub_list",
+    "Subscribed to Email Marketing": "subscribed_to_email",
     "Subscribed to List": "subscribe_list",
     "Updated Email Preferences": "update_email_preferences",
     "Dropped Email": "dropped_email",
     "Clicked SMS": "clicked_sms",
-    "Consented to Receive SMS": "consented_to_receive",
+    "Subscribed to SMS Marketing": "subscribed_to_sms",
     "Failed to Deliver SMS": "failed_to_deliver",
     "Failed to deliver Automated Response SMS": "failed_to_deliver_automated_response",
     "Received Automated Response SMS": "received_automated_response",
     "Received SMS": "received_sms",
     "Sent SMS": "sent_sms",
-    "Unsubscribed from SMS": "unsubscribed_from_sms"
+    "Unsubscribed from SMS Marketing": "unsubscribed_from_sms"
 }
 
 
@@ -72,7 +76,7 @@ class Stream(object):
         return {
             'stream': self.stream,
             'tap_stream_id': self.tap_stream_id,
-            'key_properties': [self.key_properties],
+            'key_properties': self.key_properties,
             'schema': resolved_schema,
             'metadata': self.metadata
         }
@@ -83,21 +87,21 @@ REQUIRED_CONFIG_KEYS = ["start_date"] + CREDENTIALS_KEYS
 GLOBAL_EXCLUSIONS = Stream(
     'global_exclusions',
     'global_exclusions',
-    'email',
+    ['id'],
     'FULL_TABLE'
 )
 
 LISTS = Stream(
     'lists',
     'lists',
-    'id',
+    ['id'],
     'FULL_TABLE'
 )
 
 CAMPAIGNS = Stream(
     'campaigns',
     'campaigns',
-    'id',
+    ['id'],
     'FULL_TABLE'
 )
 
@@ -126,44 +130,36 @@ def load_shared_schema_refs():
 
     return shared_schema_refs
 
-def do_sync(config, state, catalog):
-    api_key = config['api_key']
+def do_sync(config, state, catalog, headers):
     start_date = config['start_date'] if 'start_date' in config else None
-
-    stream_ids_to_sync = set()
 
     for stream in catalog.get('streams'):
         mdata = metadata.to_map(stream['metadata'])
         if metadata.get(mdata, (), 'selected'):
-            stream_ids_to_sync.add(stream['tap_stream_id'])
+            singer.write_schema(
+                stream['stream'],
+                stream['schema'],
+                stream['key_properties']
+            )
 
-    for stream in catalog['streams']:
-        if stream['tap_stream_id'] not in stream_ids_to_sync:
-            continue
-        singer.write_schema(
-            stream['stream'],
-            stream['schema'],
-            stream['key_properties']
-        )
-
-        if stream['stream'] in EVENT_MAPPINGS.values():
-            get_incremental_pull(stream, ENDPOINTS['metric'], state,
-                                 api_key, start_date)
-        else:
-            get_full_pulls(stream, ENDPOINTS[stream['stream']], api_key)
+            if stream['stream'] in EVENT_MAPPINGS.values():
+                get_incremental_pull(stream, ENDPOINTS['events'], state,
+                                    headers, start_date)
+            else:
+                get_full_pulls(stream, ENDPOINTS[stream['stream']], headers)
 
 
-def get_available_metrics(api_key):
+def get_available_metrics(headers):
     metric_streams = []
-    for response in get_all_pages('metric_list',
-                                  ENDPOINTS['metrics'], api_key):
+    for response in get_all_using_next('metric_list',
+                                  ENDPOINTS['metrics'], headers, {}):
         for metric in response.json().get('data'):
-            if metric['name'] in EVENT_MAPPINGS:
+            if metric['attributes']['name'] in EVENT_MAPPINGS:
                 metric_streams.append(
                     Stream(
-                        stream=EVENT_MAPPINGS[metric['name']],
+                        stream=EVENT_MAPPINGS[metric['attributes']['name']],
                         tap_stream_id=metric['id'],
-                        key_properties="id",
+                        key_properties=["id"],
                         replication_method='INCREMENTAL',
                         replication_keys=["timestamp"]
                     )
@@ -172,29 +168,32 @@ def get_available_metrics(api_key):
     return metric_streams
 
 
-def discover(api_key):
-    metric_streams = get_available_metrics(api_key)
+def discover(headers):
+    metric_streams = get_available_metrics(headers)
     return {"streams": [a.to_catalog_dict()
                         for a in metric_streams + FULL_STREAMS]}
 
 
-def do_discover(api_key):
-    print(json.dumps(discover(api_key), indent=2))
+def do_discover(headers):
+    print(json.dumps(discover(headers), indent=2))
 
 @singer.utils.handle_top_exception(LOGGER)
 def main():
 
     args = singer.utils.parse_args(REQUIRED_CONFIG_KEYS)
+    headers = {
+        "Authorization": f"Klaviyo-API-Key {args.config.get('api_key')}",
+        "revision": API_VERSION
+    }
 
     if args.discover:
-        do_discover(args.config['api_key'])
+        do_discover(headers)
 
     else:
-        catalog = args.catalog.to_dict() if args.catalog else discover(
-             args.config['api_key'])
+        catalog = args.catalog.to_dict() if args.catalog else discover(headers)
 
         state = args.state if args.state else {"bookmarks": {}}
-        do_sync(args.config, state, catalog)
+        do_sync(args.config, state, catalog, headers)
 
 if __name__ == '__main__':
     main()